backend/requirements.txt
flask
flask-cors
nltk
astor
numpy
scikit-learn
python-Levenshtein
